+++ actix-http/src/lib.rs
@@ -5,7 +5,7 @@
     clippy::too_many_arguments,
     clippy::new_without_default,
     clippy::borrow_interior_mutable_const,
-    clippy::write_with_newline
+//    clippy::write_with_newline
 )]
 
 #[macro_use]
@@ -16,6 +16,7 @@ mod builder;
 pub mod client;
 mod cloneable;
 mod config;
+#[cfg(feature = "compress")]
 pub mod encoding;
 mod extensions;
 mod header;